+++ src/rules/E0010.rs
@@ -1,0 +1,12 @@
+use crate::analyse::Rule;
+
+pub struct E0010 {}
+
+impl Rule for E0010 {
+    fn validate(
+        &self,
+        statement: &php_parser_rs::parser::ast::Statement,
+    ) -> Vec<crate::project::Suggestion> {
+        vec![]
+    }
+}
