+++ src/rules/mod.rs
@@ -1,6 +1,11 @@
 pub mod E001;
+pub mod E0010;
+pub mod E0011;
 pub mod E002;
 pub mod E003;
 pub mod E004;
 pub mod E005;
 pub mod E006;
+pub mod E007;
+pub mod E008;
+pub mod E009;
